+++ APHints.h
@@ -21,11 +21,11 @@ struct APHint {
 
 namespace APHints
 {
-    extern std::vector<int> HintedIDs;
+    extern std::vector<int64_t> HintedIDs;
 
     void reset();
     void handleHintMessage(const AP_HintMessage&);
-    void updateSentLocations(const std::array<uint32_t, 2> locationIDs);
+    void updateSentLocations(const std::array<int64_t, 2> locationIDs);
     void updateByItemName(const std::string itemName);
 
     void ImGuiTab();
